lib_sched.c: fix a bug when custom pin on large system with many CPUs
Because mask/cpumask is based on 'unsigned long' type, so each of it can host
64 CPUs. In this patch, cast the immediate integer '1' into type 'unsigned long'.
Without this change, '1' is treated as 32-bit signed integer, and on a sysmtem
with more than 32 CPUs, this can make pinning a task to CPU no. 32~63 fail.

This patch fix this bug: on a system with 96 cores, this command failed to pin
the two processes onto CPU 0 and 48.
  $ LMBENCH_SCHED="CUSTOM_SPREAD 0 48" lat_pipe -W 1 -N 4

Signed-off-by: Guodong Xu <guodong.xu@linaro.org>